requirements.txt
ipykernel
langchain
langchain_community
langchain_huggingface
python-dotenv
unstructured
faiss-cpu
langchain-groq
streamlit
langchain-openai
langchain_google_genai
langchain-astradb
PyPDF2